+++ lending/src/funcs/petsCreatePets.ts
@@ -5,6 +5,7 @@
 import { LendingSDKCore } from "../core.js";
 import { encodeJSON } from "../lib/encodings.js";
 import * as M from "../lib/matchers.js";
+import { compactMap } from "../lib/primitives.js";
 import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { pathToFunc } from "../lib/url.js";
@@ -52,10 +53,10 @@ export async function petsCreatePets(
 
   const path = pathToFunc("/pets")();
 
-  const headers = new Headers({
+  const headers = new Headers(compactMap({
     "Content-Type": "application/json",
     Accept: "application/json",
-  });
+  }));
 
   const context = {
     operationID: "createPets",