+++ lending/src/funcs/petsShowPetById.ts
@@ -5,6 +5,7 @@
 import { LendingSDKCore } from "../core.js";
 import { encodeSimple } from "../lib/encodings.js";
 import * as M from "../lib/matchers.js";
+import { compactMap } from "../lib/primitives.js";
 import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { pathToFunc } from "../lib/url.js";
@@ -59,9 +60,9 @@ export async function petsShowPetById(
 
   const path = pathToFunc("/pets/{petId}")(pathParams);
 
-  const headers = new Headers({
+  const headers = new Headers(compactMap({
     Accept: "application/json",
-  });
+  }));
 
   const context = {
     operationID: "showPetById",