+++ lending/RELEASES.md
@@ -1229,3 +1229,13 @@ Based on:
 - [typescript v0.7.50] lending
 ### Releases
 - [NPM v0.7.50] https://www.npmjs.com/package/ryan-lending/v/0.7.50 - lending
+
+## 2025-01-15 00:57:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.468.6 (2.493.13) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.51] lending
+### Releases
+- [NPM v0.7.51] https://www.npmjs.com/package/ryan-lending/v/0.7.51 - lending